+++ CHANGELOG.md
@@ -28,6 +28,12 @@ or major refactoring improvements.
 ### New
 
 --------------------------------------------------------------------------------
+## [0.9.0] - 2024-05-14 Fix Sutime Filter (2)
+### Fixed
+- Now keep DURATION  without 'timex-value'
+### New
+- Refactoring: SutimeTools.py holds analysis methods 
+
 ## [0.8.0] - 2024-05-06 Show Multi Mode pipeline
 ### New
 - new disp_mod in pipeline route